Fine tune and complete MessageStreamTest suite
Fine tune and finalize the MessageStreamTest suite to include all known
paths for the existing MessageStream interface and its implementations.
This change includes JavaDoc, variable/parameter/function renaming for
clarity, and method ordering to align with the MessageStream method
ordering.
More importantly, it should add tests for all methods that currently
aren't tested, as well as fixes that will follow from this.

#3063