Reduce the number of property lookups for matching entities
This commit adds a cache to the properties that have been discovered in each of the message payloads that an entity has a handler for.

This prevents lookup of properties to occur for each instance of an event that is applied to an aggregate.

With this optimization, a lookup is only required for each new payload type that is being applied.