Introduce the IndexedEventMessage
Introduce the IndexedEventMessage and a GenericIndexedEventMessage
implementation. The IndexedEventMessage will represent the combination
of a delegate EventMessage and the set of Index instances that point
towards it in the storage solution.
This implementation is paramount to move away from the
DomainEventMessage that fixated the EventMessage with the concept of a
single aggregate (through the identifier and the sequence number). The
Identifier acted as a form of index, so replacing it for a Set of
Indices allows us to support the dynamic consistency boundary more
clearly.

#3101