Rename AsyncEmbeddedEventStore to SimpleEventStore
Rename the AsyncEmbeddedEventStore to the SimpleEventStore. The biggest
difference, is that the SimpleEventStore only works for a single context
. Hence, the given context on all implemented methods is validated
against the provided context during construction. Furthermore, remove
the Clock, as that allows us to delegate all timed operations to the
EventStorageEngine, as well as solidify the storage engine's task to set
 the time of the new events (once that's in place).

#3101