Allow DLQ to handle deserialization errors
Currently the DLQ does not enqueue messages with deserialization errors, due to the following stack trace:

```
SerializedMessage.getPayload(SerializedMessage.java:81)
MessageDecorator.getPayload(MessageDecorator.java:56)
MessageDecorator.getPayload(MessageDecorator.java:56)
EventMessageDeadLetterJpaConverter.convert(EventMessageDeadLetterJpaConverter.java:59)
```

In this trace the message is first deserialized (if it wasn't yet, through the `LazyDeserializingMessage`) before being serialized again. Besides being inefficient, this causes any deserialization error to occur again during enqueuement in the DLQ, causing the event processor to go into error mode.

This PR switches the call to the `Message.serializePayload` call, which is overridden by the SerializedMessage to not deserialize/serialize if the type matches (or tries to find a converter).