Introduce the EventSink
The EventSink will pick up the base purpose of a publication spot for
events. As such, it would replace part of the EventBus, and could for
example be used by an EventGateway directly. Further, it would be a
simplification for publication implementations like AMQP and Kafka,
which only require a form of publication/sink mechanism. The EventStore
will implement the EventSink to support event publication.

#3101