Change EventCriteria and conditions into sealed interface
By adjusting the EventCriteria, AppendCondition, SourcingCondition, and
StreamingCondition into a sealed interface, we are able to better
protect the requirement of existence of the expected values. Although
the @Nonnull annotations are helpful to guide users, it does not
obstruct them from setting null. Furthermore, extensibility of the
criteria and conditions would open up the option for users to make
versions that take in nulls, while we do not want that.

#3101