Adjust the performWithBusyWaitSpin to use sleep() i.o. yield()
Adjust the performWithBusyWaitSpin to use sleep() i.o. yield(). This is
done as a test to verify if the yield() operation is ignored in certain
environments. Yield does not guarantee that anything occurs, as it's a
simple hint, while sleep is a specific command. Drawbacks are the
necessity to catch the InterruptedException inside the loop, of course.

#enhancement/replace-yield-for-short-sleep