Adjust SourcingCondition to not use Long or OptionalLong
Adjust SourcingCondition to not use Long or OptionalLong. We are
perfectly capable of defining a useful default for both
SourcingCondition#start and SourcingCondition#end, which are -1l and
Long.MAX_VALUE respectively. We do not require the API of the
SourcingCondition to provide constructors that take null values for
either, nor do we have to maintain that knowledge by returning an
OptionalLong when SourcingCondition#end is invoked.

#3101