Remove fromEntry... methods in favor of a Context supplier
To further reduce creation of an Entry by anything else than the
MessageStream, we should drop the fromEntryIterable, fromEntryStream,
fromEntryFlux, and fromFutureEntry. As the main intent for allowing an
Entry of any container was to support a solution to provide a custom
Context, we can achieve the same by providing a Function<Message,
Context>, so a contextSupplier, instead.

#3129