Cleanly await termination of the LocalSegmentAdapter
Cleanly await termination of the LocalSegmentAdapter. The current
implementation cancels all QueryProcessingTasks that are in progress
right away. This causes a pending result to get lost, even when it was
returned by the query handler. Thus, instead, we should use the
QueryProcessingTask#resultPending to validate if we are still waiting
for results within a configurable duration. This duration should be
adjustable on the AxonServerQueryBus.Builder. Only when the await
termination duration is reached should we proceed to cancelling the
queries in progress

#bug/gracefully-await-queries-in-progress