Ensure the QualifiedName is all conversion steps
Ensure the QualifiedName is all conversion steps performed between Axon
Framework and the Axon Server Connector Java. As we need to support the
"old" API of Axon Server, there will always be a solution that does not
have a given QualifiedName. Hence, it should deduce it from the existing
 API, for which a decision should be made. This work is captured in
 TODOs.
Ensure the QualifiedName is present in LazyIdentifierDomainEventMessage.
 Note that this does not suffice for the AnnotatedAggregate entirely, as
  we need to rethink the apply method. However, the chance is high this
  operation will be omitted entirely, as the
  EventBus/EventStoreTransaction will be passed along the
  ProcessingContext for message handlers to retrieve.

#3085