Adjust ResourceKey to only have withLabel operation
Adjust ResourceKey to only have withLabel operation. Thus, each
invocation of ResourceKey#withLabel generates a new label. The label in
turn provides a, well, label, that clarifies what it's for. If a
resource should be shared among interested parties, the ResourceKey
should simply be shared. And, by enforcing uniqueness through the
construction, we ascertain there will never be a clash with users
constructing similarly labeled keys.

#3185