Rework the MessageHandler interface layering entirely
Instead of imposing a handle method on the MessageHandler interface,
simply change it into a sealed interface that only permits a
CommandHandler, EventHandler, and QueryHandler. Furthermore, do the
same for the MessageHandlerRegistry. This means we introduce dedicated
command, event, and query handler registries. Lastly, let the handling
components implement their respective specific handler and specific
registry instead of the generic versions. All this should ease typing
and provide exact methods for subscribe, handle, and supported messages.

#3067