Correctly construct AppendCondition from source calls (part 2)
This PR introduces the implementation and tests for the AsyncEventSourcingRepository to ensure that events are appended using a correct AppendCondition, which is constructed using the SourcingConditions from each of the source calls on that repository in the same ProcessingContext.

The ProcessingContext parameter has been removed from the methods on an EventStoreTransaction, as the context is already provided on retrieval of that transaction. The transaction itself is already aware of the ProcessingContext it operates under.