Make EventHandler return a MessageStream.Empty
Make the EventHandler interface return a MessageStream.Empty instead of
MessageStream<NoMessage>. In doing so, we clarify that an event handler
does not return anything. Note that this change causes problems for
returning a failed message stream. This is currently resolved by having
a dedicated FailedEmptyMessageStream implementation, which can be
constructed with the static MessageStream.Empty.failed operation.
Lastly, we can remove the NoMessage

#3067