Make CommandHandler return a MessageStream.Single
Make the CommandHandler interface return a MessageStream.Single instead
of MessageStream<? extends CommandResultMessage<?>>. In doing
so, we clarify that a command handler returns a singular result at all
times. Note that this change causes problems for returning a failed
message stream. This is currently resolved by having
a dedicated FailedSingleMessageStream implementation, which can be
constructed with the static MessageStream.Single.failed operation.
Furthermore, the usages of the CommandHandler interface currently break
when message stream operations are chained, as occurs in the
InterceptingCommandBus and AggregateAnnotationCommandHandler. This is
resolved by adjusting the MessageStream.cast method to cast the entire
stream instead of only the generic message. This brings the
implementation into a partially working state atm.

#3067