Fix infinite event source call and missing ConsistencyMarker
The AsyncInMemoryEventStorageEngine does not call the availability callback when starting a sourcing stream and the store is empty.
This causes the stream to hang indefinitely when using normal `MessageStream` methods like `reduce` (which the `EventSourcingAsyncRepository` does).

To fix this, we now call the availability callback immediately when the store is empty.
For flux this worked, because it called `hasNextAvailable`.

Additionally, the `Context` of the `EventMessage`s were missing the `ConsistencyMarker`, which caused always cause the marker to be origin, and any event to be appended a conflict when used in conjunction with the `EventSourcingAsyncRepository`.