Prevent ORIGIN marker when source event stream is empty
In the current implementation of the DefaultEventStoreTransaction the marker for the commit is always the lowest bound of the two.
This works correctly if the two streams both return at least one event. However, if one stream returns no events, which means there is no matching event up to HEAD, the lowest of the two will be ORIGIN (-1) and always lead to a conflict.
This makes it impossible to load two models in the same transaction, of which one might not match at least one event.

This PR will keep the current ConsistencyMarker in place if the stream is empty. Whenever the entire stream is read, resulting in no events, the lower bound is automatically the current marker.