Fix failing Spring Boot tests due to 4.11.x merge.
First, the HttpMessageConvertersAutoConfiguration autoconfiguration class requires a singular ObjectMapper bean because it imports JacksonHttpMessageConvertersConfiguration. It breaks, as there *are* two ObjectMapper implementations in the bean registry. As such, I have excluded this for a few tests.

Second and last, the assertion for the number of interceptors in the EventProcessingModuleConfigTest was wrong, as there is an additional interceptor added by the AxonTimeoutConfigurerModule introduced in 4.11. I have disabled the module adding this interceptor through a property.