Introduce a RootConfigurer, DefaultRootConfigurer, and RootConfiguration
Introduce a RootConfigurer, DefaultRootConfigurer, and RootConfiguration
The RootConfigurer interface serves the purpose of being the first
Configurer object to be created by any Axon app. As such, it provides
static factory methods to instantiate the initial config. Furthermore,
the interface allows for adjusting lifecycle phase timeouts, the
building of a Configuration, and starting. Furthermore, the
DefaultRootConfigurer is the implementation that constructs the
RootConfigurationImpl. It is this RootConfigurationImpl that does the
start and shutdown handler invocation, in an identical way as the
DefaultConfigurer does.

#3067