Implement annotation-based resolving of models
Models can no be resolved by injecting parameters of a `CommandHandler` with `@InjectModel`. By default, it will use an annotation-based `IdResolver` to resolve the identifier of the model based on the `@TargetModelIdentifier`, before resolving it from the `ModelRegistry`

However, it can also be customized. Users can specify a specific property as the identifier.

Lastly, they can implement their own `IdResolver` to make it as complicated as they want.

Several `ParameterResolver`s have been introduced to make this work.

Unfortunately, compound models in the current code base do not work, as OR tags are currently not supported, nor an append when sourced from multiple tags. However, I left the disabled test in to showcase how it would work. I confirmed through manual code edits that this test works. See `CompoundModelIdentifierCommandHandlingComponentTest`.