Proposal for the `EventCriteria` building
Right now, the `EventCriteria` cannot do an OR condition. During a discussion, Mateusz and I came up with the following idea to build the EventCriteria flexibly.

With this new approach, users can potentially build a condition as complex as they want, while also being very simply for the more simple use-cases.

There would be no more builder, but the user can build the EventCriteria through a series of methods.