Break RootConfigurerTest into a test suite
Break RootConfigurerTest into a test suite. Since each module is also a
Configurer, it is beneficial to have a ConfigurerTestSuite. However, as
Modules typically require the LifecycleSupportingConfiguration, a
ModuleTestSuite should also be present with its own test subject
creation method. Lastly, we should provide a test suite implementation
for the RootConfigurer and InfraConfigurer

#3067