Only create for exact AggregateNotFoundExceptions
Only create for exact AggregateNotFoundExceptions. Otherwise the
LockingRepository might accidentally create an aggregate when an
AggregateDeletedException was thrown. The AggregateDeletedException
means the event stream for this aggregate still exists, but the user
marked it as closed. Since AggregateDeletedException is an
implementation of the AggregateNotFoundException, the LockingRepository
would create a new instance. Although ideally we'd check the
AggregateDeletedException directly in the LockingRepository, since a
deleted aggregate is an event sourcing specific, and the locking
repository is not, we cannot do this.

#3323