Introduce shift between ApplicationConfigurer and Module building
Introduce shift between ApplicationConfigurer and Module building, to
clarify that a Module is built with a parent Configuration, and a base
Configurer is not. This enforces removal of the Configurer#build method,
 which is not part of a new interface called ApplicationConfigurer. This
 new interface also contains the start operation, providing a common
 start tasks between the AxonApplication and MessagingConfigurer.
 Furthermore, the Module now has a build(Configuration) operation. This
method returns the resulting module's configuration (without a name for
the time being). This shift allows us to remove the constructor on the
AbstractConfigurer, but enforces us to provide a means to set the parent
 configuration on the AbstractConfigurer. On top of that, the common
 build tasks, like enhancing and module building, should be reused in
 any ApplicationConfigurer AND Module implementation. The Module might
 benefit from a common abstract class, but I felt that work redundant
 for a first draft. Lastly, to enforce building of a Module during the
 ApplicationConfigurer#build, we should remove the ModuleBuilder
 entirely.

#3067