Introduce an AnnotatedEventSourcedEntityBuilder
Introduce an AnnotatedEventSourcedEntityBuilder. This builder replaces
the AnnotationBasedEventSourcingEntityRepository, as that essentially
was an entity builder. The applicable, non-duplicate tests from the
AnnotationBasedEventSourcingEntityRepositoryTest have been moved and
rewritten in the AnnotatedEventSourcedEntityBuilderTest. Furthermore, I
have pushed the EventSourcedEntityFactory into the
AsyncEventSourcingRepository as a first class citizen. This eliminates
the vague lambda in there. While touching the
AsyncEventSourcingRepository, I've made some minor improvements to align
 with other recently introduced components.

#3345