Allow hierarchic configuration of ParameterResolverFactory and StateManager
Introduces the HierarchicalParameterResolverFactory and HierarchicalStateManager and configuration defaults for them, so that the module hierarchy is taken into account when resolving parameters and entities.

Both the BaseModule and MessagingConfigurer configure the ParameterResolverFactoryConfigurationDefaults to have at least the configuration and classpath parameter resolvers.
The ModellingConfigurer and the DefaultStatefulCommandHandlingModule use the StateManagerConfigurationDefaults to configure hierarchical state manager, as well as a parameter resolver to that level of configuration for that level's state manager.

Due to a circular reference of the StateManagerConfigurationDefaults, I had to change the constructor of the InjectEntityParameterResolver