Introduce lambda-based EventStateAppliers in EventSourcedEntityBuilder
Introduces 4 new functions on the `EventSourcedEntityBuilder` to allow more flexibility in registering event-sourcing handlers.
There is a variant with `QualifiedName` and one without. The latter uses the `MessageTypeResolver` so the user doesn't have to specify it.
Both of these variants have a `BiFunction` and a `BiConsumer` variant, so the user can decide whether he/she wants to return the model or not. This is useful for, for example, immutable models.

To facilitate this, the `SingleEventEventStateApplier` has been born, which takes in the message type, payload type, and a function, that will only be called if it matches.

Last but not least, I have changed the builder to allow registration of many `EventStateAppliers` and event-sourcing handlers. These are combined into a `MultiEventStateApplier` upon build.