Revamp EventStateApplier into EntityEvolver
The EventStateApplier does not align with our current line of thinking
what this function does. It should, instead, Evolve a given Entity (of
type E), returning the evolved solution. This merits a rename of the
functional interface (now EntityEvolver), the method (now evolve), the
parameter (now entity i.o. model), and the generic (now E i.o. M). As
part of this shift, we should align the newly introduced entity evolvers
 to be more to the point the SingleEventTypeStateApplier becomes the
 PayloadConvertingEntityEvolver, as it enforces the format to be the
 given payload. Furthermore, it should not filter at all, as this should
 be done by the given lambda or higher up. The MultiEventStateApplier
 becomes the SimpleEntityEvolvingComponent, implementing a new
 EntityEvolvingComponent interface. This simple implementation has a
 collection of EntityEvolvers per QualifiedName. As such, it filters
 based on the qualified name in the given event. Furthermore, as an
 EntityEvolvingComponent, it has a set of supported events. Lastly, the
 builder should be cleaned up slightly.

#3286