Introduce definitions to construct annotation-related classes
The `@EventCriteriaBuilder` methods need a `MessageTypeResolver` to receive from the AnnotationBasedEventCriteriaResolver. However, we don't want to pass the entire configuration into such a component, allowing the user to do all sorts of processes to build the criteria. These methods should remain simple.

As a compromise, the AnnotationBasedEventCriteriaResolverDefinition has been introduced, defining the arguments that can be used for the construction of such a class. This allows us to add specific component to the contruction, without relying on the entire configuration, or a ton of reflection to determine arguments.

The EventSourcedEntityFactory has received the same treatment to align with the approach. This will allow us the same flexibility in the future.