Adjust interceptor behavior for child entities
Adjust interceptor behavior for child entities. Instead of preemptively
checking the matching interceptors, simply push all interceptors into a
ChainedMessageHandlerInterceptorMember. Doing so, we ensure the
InterceptorChain is present as a ThreadLocal, allowing the
InterceptorChainParameterResolver to inject it on time. If there are no
interceptors, use the NoMoreInterceptors instance. This shifts makes it
so that child entities behave the same as the parent aggregate when it
comes to interceptor behavior.

#3385