Add allocation-size details
Add allocation size details to the warning and sample. If users set a
pooled optimizer, they would prepare a (by default) set of 50 global
indices shared over several instances. This will definitely incur a lot
of gaps, causing performance degradation of Axon's JPA RDBMS solution.
As a solution, users can adjust the allocation-size of the sequence
generator to one. Although this does not entirely resolve the issue, it
makes the window of opportunity a lot smaller.

#3374