Remove `? extends` from many MessageStream results.
I experienced problems while creating tests for command handling components. It was very hard to get components to capture the generic type correctly due to the `? extends`.

That's when I wondered if we couldn't remove it. It definitely would look a lot more clear on the interfaces for the user.

Turns out we can. I am making this PR as a proposal. Perhaps I missed something obvious, the reason why the wildcard was there, but if we can remove it, I think we should.