Adjust SourcingMessageStream to peek the next entry for the marker
Instead of storing an AtomicReference of the ConsistencyMarker, which
can cause accidental inconsistent reads midway through sourcing a
stream, we should ResultStream#peek the next entry upon a
MessageStream#next invocation. To that end, the SourcingMessageStream
should maintain an AtomicReference of the previously read entry. From
there, we can ResultStream#peek the next entry. If the next entry has a
consistency marker, we reached the end. If the next entry does not, we
are simply consuming entries. By replacing the previous entry with the
current entry as the first step, we ensure that subsequent invocations
take the previous peeked response.

#3239