Creational command handlers for (immutable) entities
This PR is an addition on PR https://github.com/AxonFramework/AxonFramework/pull/3418 (Declarable entity modeling).

One of the annoyances in Axon Framework 4 right now is that every property of an entity/aggregate has to be mutable, due to the required existence of a no-arg constructor to initialize the state. This requires `lateinit var` variables, or nullable variables, to work.

To fix the nullability, users can now declare a constructor that takes the first `EventMessage`, or its payload. Users can still opt to go for no-arg constructors to. As we cannot know if the event was consumed or not, regular event-sourcing handlers that match will be invoked as well. Users can also do this programmatically through the `EventSourcedEntityFactory`, defining separate factories for new instances without prior events, and existing instances with the first event message. By default the first event message is ignored for backwards compatability.

This PR also introduces the ability to define a creational command handler on the `EntityModel`. This is automatically invoked when there is no entity present during the invocation of the `EntityCommandHandlingComponent`. This behavior can be adjusted through the introduced `CreationPolicy`.

A creational command handler does not return an entity. It has the responsibility to validate the command and create the entity. When using event sourcing, this is publishing an event, as publishing this event will automatically be noticed by the Repository, and the entity will be created through the configured `EventSourcedEntityFactory` with said (optional) first event.

 A creational and a regular (aka instance) command handler can be present for the same command type at the same time, to be able to handle both situations.

 This PR does not include the required functionality to do this using Reflection. This will be added in a future PR.