Adjust source operation to return a Consistency Marker entry at the end
Adjust the AggregateBasedJpaEventStorageEngine to return a single
AggregateBasedConsistencyMarker. The complexity lies in the fact that
the JPA-based event store retrieves aggregate streams, which we
concatenate with one another. When each of these aggregate streams
complete, we need to capture an AggregateBasedConsistencyMarker. This
marker, together with the aggregate's sourced events, form the response
of a single aggregate sourcing operation. These results should then be
combined again. Once all streams have been read, we need to attach an
AggregateBasedConsistencyMarker based on all aggregate streams that have
 been read.

#3239