Adjust MetaData to a Map of String values
Adjust MetaData to a Map of String values. Although allowing a wildcard
as the value is friendly to the user, it abstracts a way the fact that
metadata is stored and sent over the wire. By enforcing the value to a
String, we greatly simplify our side of the coin. A follow-up commit/PR
will contain get and put operations that accept a Converter. Users that
do want more complex objects in MetaData (which we thus not recommend)
can use these.

#3423