Make all components use the ProcessingContext
All (legacy) components now pass on the `ProcessingContext` into all subcomponents that might need it. Reason for this being necessary is the removal of `Message` on the `ParameterResolver` interface so it can be used for reflection outside of message handling as well.

The LegacyUnitOfWork now creates a LegacyMessageSupportingContext and passes this into the execute function, that can then be passed on by components.

Most changes in this PR is passing through the parameter, and then adjusting the tests.

However, the NoProcessingContext has been modified into the SimpleProcessingContext. A processing context that rejects resources is far from useful in any situation, as components may pass that along and expect resources to be in there. This context does not register lifecycle handlers.

As next step we might consider creating a MessageProcessingContext that extends ProcessingContext and exposes a `getMessage()` instead. This would make instanceof available.
For now, I decided against doing this. We can do this in a future PR.