Deprecate aggregate versioning support
Deprecate aggregate versioning support. Similarly as with Conflict
Resolution, aggregate versioning has been used extremely little since
AxonIQ's inception in 2017. Although it has been around since version 1
.2 of Axon Framework, we deem it unnecessary to construct a renewed
version of this support for AF5. What's important in this decision, is
that fact that Axon Framework will no longer dictate the aggregate
sequence numbers. This is in line with the Event Store shift to support
DCB, wherein the overall global index of the events are used as ordering
. Differently put, a second ordering per aggregate isn't necessary.
Furthermore, if an event would belong to several tags/aggregates, that
would require us to define several aggregate sequence numbers. Lastly,
this commit **does not** remove the functionality outright, as we view
it as part of the old Aggregate basics. As those will be moved mostly as
 is to the legacy/migration package, marking it as deprecated is
 sufficient.

#3433