Split ChildEntityMatcher for clearer message forwarding
The current EntityModel uses a ChildEntityMatcher to determine both to which targets to forward a command, but as well to select a command.

This feels strange in the API, as they have inherently different requirements. For events this has a 0, 1 or many result, and for commands it has 1 or 0.

As such, this PR splits this into the CommandTargetResolver, which requires an instance to be returned, and the EventTargetMatcher, which is a predicate taking in the message, candidate and processing context.