Rename EventSourcedComponent to EntityEvolvingComponent
Right now, the `EventSourcedComponent` is part of the `eventsourcing` package. This is a component that represents a set of `EntityEvolvers`.

This concept, however, is not unique to event-sourcing.
State-stored entities can also be evolved by declaring `@EventHandler`-annotated methods, after which the new state can be saved.

As such, this PR renames the `EventSourcedComponent` to `EntityEvolvingComponent` and moves it to the `modelling` module. Some related classes have received similar treatment.