Rename tail to first and head to latest
The use of head and tail for the position at the END of a message
stream, and tail for the START of a message stream, causes a lot of
confusion to a lot of people. We would be far better off by renaming
these to what they resemble: the first position known in the message
stream, and the current latest position of that potentially infinite
stream. Given that the current new implementation already made mistakes
in this area is proof enough it should be changed.