Rename EntityModel to EntityMessagingMetamodel across the codebase for consistency.
This PR renames the `EntityModel` to `EntityMessagingMetamodel` and gives related classes a similar treatment.

When we talk about a model in AF, we talk about the user's entities and how they are modeled together. Our model is the meta-model, describing the messaging requirements and capabilities of that model.

As such, the following renames have been done:
- `EntityModel` to `EntityMessagingMetamodel`
- `AnnotatedEntityModel` to `AnnotatedEntityMessagingMetamodel`
- `SimpleEntityModel` to `ConcreteEntityMessagingMetamodel`
- `PolymorphicEntityModel` to `PolymorphicEntityMessagingMetamodel`
- `EntityChildModel` to `EntityChildMessagingMetamodel`
- `ListEntityChildModel` to `ListEntityChildMessagingMetamodel`
- `SingleEntityChildModel` to `SingleEntityChildMessagingMetamodel`
- `AbstractEntityChildModel` to `AbstractEntityChildMessagingMetamodel`