Adjust EventGateway, EventSink, and EventStore API
## EventGateway
The EventGateway should have a nullable ProcessingContext. This can
either be used to publish events within the existing
context/transaction, or to add correlation data.
Furthermore, change the return type to a CompletableFuture of void. In
doing so, the EventGateway aligns with our async focus.
By doing this, the EventGateway aligns neatly with the CommandGateway,
simplifying the process for our users.

## EventSink
Make the ProcessingContext nullable, just as with the EventGateway and
CommandBus. Furthermore, combine the publish without context and with
context into a single method, returning a CompletableFuture of void.
By doing so, the EventSink more clearly aligns with the CommandBus.

## EventStore & SimpleEventStore
Make the ProcessingContext nullable, just as with the EventGateway and
EventSink. Added, remove the overridden publish operations and instead
hold these on the SimpleEventStore. Since we aren't certain of a
ProcessingContext but do NEED one of these, we cannot provide a default
publish implementation. Instead, let the SimpleEventStore construct a
UnitOfWork. Ideally this logic resides with a SimpleEventSink. But,
while awaiting #3392, this is the best we can do.

## Rest
Fix all usages and tests according to these changes.

#3470