Move config defaults into separate test and expand
Move config defaults into separate test and expand. This shows that the
ConfigurationEnhancer disablement through a bean requires a minor tweak
in the private invokeEnhancers() operation. While using a stream, the
disabledEnhancers cannot be updated as needed. Hence, move that into a
separate for-loop with if-check inside for the disabled entries. This
change should also be made in the DefaultComponentRegistry and be
documentated.

#3075