Introduce SearchScope for hasComponent
Introduce SearchScope for hasComponent. This shift is mandatory to
ensure ConfigurationEnhancers can choose to check if a component is
present in the current ComponentRegistry or present **anywhere**. This
is in turn required, as some enhancers should act on every registry,
while others should act on the entire configuration as a whole.

#3075