fix: `AxonTestFixture` works with `SpringAxonApplication` (#3512)
Usage example:
```kotlin
@SpringBootTest
internal class BuildDwellingSpringTest @Autowired constructor(configurer: ApplicationConfigurer) {
    
    private val fixture: AxonTestFixture = AxonTestFixture.with(configurer)

    @Test
    fun `given not built dwelling, when build, then built`() {
        val dwellingId = UUID.randomUUID().toString()
        val creatureId = "angel"
        val costPerTroop = mapOf(ResourceType.GOLD to 3000, ResourceType.GEMS to 1)

        fixture.given()
            .noPriorActivity()
            .`when`()
            .command(BuildDwelling(dwellingId, creatureId, costPerTroop))
            .then()
            .events(
                DwellingBuilt(
                    dwellingId = dwellingId,
                    creatureId = creatureId,
                    costPerTroop = costPerTroop
                )
            )
    }

    @TestConfiguration
    class TestConfig {

        @Bean
        fun recordingEnhancer() = MessagesRecordingConfigurationEnhancer()
    }

}
```

If you run this test without declaring the enhancer, the test will fail
with the following error:
```
org.axonframework.test.FixtureExecutionException: CommandBus is not a RecordingCommandBus. This may happen in Spring environments where the MessagesRecordingConfigurationEnhancer is not properly registered. Please declare MessagesRecordingConfigurationEnhancer as a bean in your test context. Note: This configuration may be subject to change until the 5.0.0 release.
```

I think later we could introduce some annotation like
`@AxonSpringBootTest` which will inject the `AxonTestFixture` but we
need to think about it carefully, so this workaround should be enough as
a fix.

Recording classes are public now, if someone wants to experiment with
it/make own implementations, but we don't guarantee the compatibility in
the future, so I've marked them as `@Internal`.