Convert payload before extracting id from message for AnnotatedEntityMetamodel
The current default configuration of the @EventSourcedEntity annotation is to scan the payload of a message for an annotated field. However, when using a distributed bus, this payload is serialized, and only the AnnotatedEntityMetamodel knows the representation that will be expected.

As such, this PR introduced the AnnotatedEntityMetamodelEntityIdResolver and sets it as the default. This will still scan for annotations, but also wraps the AnnotationBasedEntityIdResolver with a decorator that converts the payload before scanning it. That way, it doesn't scan a `byte[]`, but a java representation.