Fix interruption handling in message handling
The 4.11-introduced timeouts for message handling and the UnitOfWork showed conflicting behavior. This PR ensures the following three situations are handled well:

1. The message handler takes too long
2. The UnitOfWork takes too long
3. Some unrelated interrupt is done, for example during shutdown of VM

By checking the Thread.interrupted() flag, the (optional) thrown exception, and the interrupted status of the tasks, we can now distinguish these situations.

Examples:
- If a Saga's CommandGateway.sendAndWait is interrupted (causing a Thead.interrupt(), but no exception), this is detected and a `TimeoutException` is thrown.
- If the batch timeout is exceeded, the message-level task will no longer interrupt, it will be the UoW level task. This is ignored by the lower level, but the lower level will cancel its interrupt schedule

It's quite a bit of complicated logic. Please see the AxonTimeLimitedTask for proper documentation in the javadoc.